backend/requirements.txt
requests
beautifulsoup4
pandas
fake_useragent
ddgs
python-dotenv

streamlit
plotly
fastapi
uvicorn
python-multipart